Replace a given separator across dataframe

Takes a given separator and replaces it with an alternative across the dataframe.

replace_sep(data, old_sep, new_sep = "; ")

Arguments

data
A dataframe containing rows of systematic review data.

old_sep
An existing character used to separate values within the data. The default is set to ';'.

new_sep
A replacement character used to separate values within the data. The default is set to ';'.

Examples

data <- data %>%
    replace_sep(old_sep = 'xxx', new_sep = '; ');

#> Error in (function (..., row.names = NULL, check.rows = FALSE, check.names = TRUE,     fix.empty.names = TRUE, stringsAsFactors = default.stringsAsFactors()) {    data.row.names <- if (check.rows && is.null(row.names))         function(current, new, i) {            if (is.character(current))                 new <- as.character(new)            if (is.character(new))                 current <- as.character(current)            if (anyDuplicated(new))                 return(current)            if (is.null(current))                 return(new)            if (all(current == new) || all(current == ""))                 return(new)            stop(gettextf("mismatch of row names in arguments of 'data.frame', item %d",                 i), domain = NA)        }    else function(current, new, i) {        if (is.null(current)) {            if (anyDuplicated(new)) {                warning(gettextf("some row.names duplicated: %s --> row.names NOT used",                   paste(which(duplicated(new)), collapse = ",")),                   domain = NA)                current            }            else new        }        else current    }    object <- as.list(substitute(list(...)))[-1L]    mirn <- missing(row.names)    mrn <- is.null(row.names)    x <- list(...)    n <- length(x)    if (n < 1L) {        if (!mrn) {            if (is.object(row.names) || !is.integer(row.names))                 row.names <- as.character(row.names)            if (anyNA(row.names))                 stop("row names contain missing values")            if (anyDuplicated(row.names))                 stop(gettextf("duplicate row.names: %s", paste(unique(row.names[duplicated(row.names)]),                   collapse = ", ")), domain = NA)        }        else row.names <- integer()        return(structure(list(), names = character(), row.names = row.names,             class = "data.frame"))    }    vnames <- names(x)    if (length(vnames) != n)         vnames <- character(n)    no.vn <- !nzchar(vnames)    vlist <- vnames <- as.list(vnames)    nrows <- ncols <- integer(n)    for (i in seq_len(n)) {        xi <- if (is.character(x[[i]]) || is.list(x[[i]]))             as.data.frame(x[[i]], optional = TRUE, stringsAsFactors = stringsAsFactors)        else as.data.frame(x[[i]], optional = TRUE)        nrows[i] <- .row_names_info(xi)        ncols[i] <- length(xi)        namesi <- names(xi)        if (ncols[i] > 1L) {            if (length(namesi) == 0L)                 namesi <- seq_len(ncols[i])            vnames[[i]] <- if (no.vn[i])                 namesi            else paste(vnames[[i]], namesi, sep = ".")        }        else if (length(namesi)) {            vnames[[i]] <- namesi        }        else if (fix.empty.names && no.vn[[i]]) {            tmpname <- deparse(object[[i]], nlines = 1L)[1L]            if (substr(tmpname, 1L, 2L) == "I(") {                ntmpn <- nchar(tmpname, "c")                if (substr(tmpname, ntmpn, ntmpn) == ")")                   tmpname <- substr(tmpname, 3L, ntmpn - 1L)            }            vnames[[i]] <- tmpname        }        if (mirn && nrows[i] > 0L) {            rowsi <- attr(xi, "row.names")            if (any(nzchar(rowsi)))                 row.names <- data.row.names(row.names, rowsi,                   i)        }        nrows[i] <- abs(nrows[i])        vlist[[i]] <- xi    }    nr <- max(nrows)    for (i in seq_len(n)[nrows < nr]) {        xi <- vlist[[i]]        if (nrows[i] > 0L && (nr%%nrows[i] == 0L)) {            xi <- unclass(xi)            fixed <- TRUE            for (j in seq_along(xi)) {                xi1 <- xi[[j]]                if (is.vector(xi1) || is.factor(xi1))                   xi[[j]] <- rep(xi1, length.out = nr)                else if (is.character(xi1) && inherits(xi1, "AsIs"))                   xi[[j]] <- structure(rep(xi1, length.out = nr),                     class = class(xi1))                else if (inherits(xi1, "Date") || inherits(xi1,                   "POSIXct"))                   xi[[j]] <- rep(xi1, length.out = nr)                else {                  fixed <- FALSE                  break                }            }            if (fixed) {                vlist[[i]] <- xi                next            }        }        stop(gettextf("arguments imply differing number of rows: %s",             paste(unique(nrows), collapse = ", ")), domain = NA)    }    value <- unlist(vlist, recursive = FALSE, use.names = FALSE)    vnames <- unlist(vnames[ncols > 0L])    if (fix.empty.names && any(noname <- !nzchar(vnames)))         vnames[noname] <- paste0("Var.", seq_along(vnames))[noname]    if (check.names) {        if (fix.empty.names)             vnames <- make.names(vnames, unique = TRUE)        else {            nz <- nzchar(vnames)            vnames[nz] <- make.names(vnames[nz], unique = TRUE)        }    }    names(value) <- vnames    if (!mrn) {        if (length(row.names) == 1L && nr != 1L) {            if (is.character(row.names))                 row.names <- match(row.names, vnames, 0L)            if (length(row.names) != 1L || row.names < 1L ||                 row.names > length(vnames))                 stop("'row.names' should specify one of the variables")            i <- row.names            row.names <- value[[i]]            value <- value[-i]        }        else if (!is.null(row.names) && length(row.names) !=             nr)             stop("row names supplied are of the wrong length")    }    else if (!is.null(row.names) && length(row.names) != nr) {        warning("row names were found from a short variable and have been discarded")        row.names <- NULL    }    class(value) <- "data.frame"    if (is.null(row.names))         attr(value, "row.names") <- .set_row_names(nr)    else {        if (is.object(row.names) || !is.integer(row.names))             row.names <- as.character(row.names)        if (anyNA(row.names))             stop("row names contain missing values")        if (anyDuplicated(row.names))             stop(gettextf("duplicate row.names: %s", paste(unique(row.names[duplicated(row.names)]),                 collapse = ", ")), domain = NA)        row.names(value) <- row.names    }    value})(... = "", list = "character", package = character(0), lib.loc = character(0),     verbose = c("getOption", "verbose"), envir = ".GlobalEnv",     overwrite = "TRUE", c("{", "fileExt <- function(x) {\n    db <- grepl(\"\\\\.[^.]+\\\\.(gz|bz2|xz)$\", x)\n    ans <- sub(\".*\\\\.\", \"\", x)\n    ans[db] <- sub(\".*\\\\.([^.]+\\\\.)(gz|bz2|xz)$\", \"\\\\1\\\\2\", x[db])\n    ans\n}",     "names <- c(as.character(substitute(list(...))[-1]), list)",     "if (!is.null(package)) {\n    if (!is.character(package)) \n        stop(\"'package' must be a character string or NULL\")\n    if (any(package %in% \"base\")) \n        warning(\"datasets have been moved from package 'base' to package 'datasets'\")\n    if (any(package %in% \"stats\")) \n        warning(\"datasets have been moved from package 'stats' to package 'datasets'\")\n    package[package %in% c(\"base\", \"stats\")] <- \"datasets\"\n}",     "paths <- find.package(package, lib.loc, verbose = verbose)",     "if (is.null(lib.loc)) paths <- c(path.package(package, TRUE), if (!length(package)) getwd(), paths)",     "paths <- unique(normalizePath(paths[file.exists(paths)]))",     "paths <- paths[dir.exists(file.path(paths, \"data\"))]",     "dataExts <- tools:::.make_file_exts(\"data\")", "if (length(names) == 0) {\n    db <- matrix(character(), nrow = 0, ncol = 4)\n    for (path in paths) {\n        entries <- NULL\n        packageName <- if (file_test(\"-f\", file.path(path, \"DESCRIPTION\"))) \n            basename(path)\n        else \".\"\n        if (file_test(\"-f\", INDEX <- file.path(path, \"Meta\", \"data.rds\"))) {\n            entries <- readRDS(INDEX)\n        }\n        else {\n            dataDir <- file.path(path, \"data\")\n            entries <- tools::list_files_with_type(dataDir, \"data\")\n            if (length(entries)) {\n                entries <- unique(tools::file_path_sans_ext(basename(entries)))\n                entries <- cbind(entries, \"\")\n            }\n        }\n        if (NROW(entries)) {\n            if (is.matrix(entries) && ncol(entries) == 2) \n                db <- rbind(db, cbind(packageName, dirname(path), entries))\n            else warning(gettextf(\"data index for package %s is invalid and will be ignored\", sQuote(packageName)), domain = NA, call. = FALSE)\n        }\n    }\n    colnames(db) <- c(\"Package\", \"LibPath\", \"Item\", \"Title\")\n    footer <- if (missing(package)) \n        paste0(\"Use \", sQuote(paste(\"data(package =\", \".packages(all.available = TRUE))\")), \"\\n\", \"to list the data sets in all *available* packages.\")\n    else NULL\n    y <- list(title = \"Data sets\", header = NULL, results = db, footer = footer)\n    class(y) <- \"packageIQR\"\n    return(y)\n}",     "paths <- file.path(paths, \"data\")", "for (name in names) {\n    found <- FALSE\n    for (p in paths) {\n        tmp_env <- if (overwrite) \n            envir\n        else new.env()\n        if (file_test(\"-f\", file.path(p, \"Rdata.rds\"))) {\n            rds <- readRDS(file.path(p, \"Rdata.rds\"))\n            if (name %in% names(rds)) {\n                found <- TRUE\n                if (verbose) \n                  message(sprintf(\"name=%s:\\t found in Rdata.rds\", name), domain = NA)\n                thispkg <- sub(\".*/([^/]*)/data$\", \"\\\\1\", p)\n                thispkg <- sub(\"_.*$\", \"\", thispkg)\n                thispkg <- paste0(\"package:\", thispkg)\n                objs <- rds[[name]]\n                lazyLoad(file.path(p, \"Rdata\"), envir = tmp_env, filter = function(x) x %in% objs)\n                break\n            }\n            else if (verbose) \n                message(sprintf(\"name=%s:\\t NOT found in names() of Rdata.rds, i.e.,\\n\\t%s\\n\", name, paste(names(rds), collapse = \",\")), domain = NA)\n        }\n        if (file_test(\"-f\", file.path(p, \"Rdata.zip\"))) {\n            warning(\"zipped data found for package \", sQuote(basename(dirname(p))), \".\\nThat is defunct, so please re-install the package.\", domain = NA)\n            if (file_test(\"-f\", fp <- file.path(p, \"filelist\"))) \n                files <- file.path(p, scan(fp, what = \"\", quiet = TRUE))\n            else {\n                warning(gettextf(\"file 'filelist' is missing for directory %s\", sQuote(p)), domain = NA)\n                next\n            }\n        }\n        else {\n            files <- list.files(p, full.names = TRUE)\n        }\n        files <- files[grep(name, files, fixed = TRUE)]\n        if (length(files) > 1) {\n            o <- match(fileExt(files), dataExts, nomatch = 100)\n            paths0 <- dirname(files)\n            paths0 <- factor(paths0, levels = unique(paths0))\n            files <- files[order(paths0, o)]\n        }\n        if (length(files)) {\n            for (file in files) {\n                if (verbose) \n                  message(\"name=\", name, \":\\t file= ...\", .Platform$file.sep, basename(file), \"::\\t\", appendLF = FALSE, domain = NA)\n                ext <- fileExt(file)\n                if (basename(file) != paste0(name, \".\", ext)) \n                  found <- FALSE\n                else {\n                  found <- TRUE\n                  zfile <- file\n                  zipname <- file.path(dirname(file), \"Rdata.zip\")\n                  if (file.exists(zipname)) {\n                    Rdatadir <- tempfile(\"Rdata\")\n                    dir.create(Rdatadir, showWarnings = FALSE)\n                    topic <- basename(file)\n                    rc <- .External(C_unzip, zipname, topic, Rdatadir, FALSE, TRUE, FALSE, FALSE)\n                    if (rc == 0) \n                      zfile <- file.path(Rdatadir, topic)\n                  }\n                  if (zfile != file) \n                    on.exit(unlink(zfile))\n                  switch(ext, R = , r = {\n                    library(\"utils\")\n                    sys.source(zfile, chdir = TRUE, envir = tmp_env)\n                  }, RData = , rdata = , rda = load(zfile, envir = tmp_env), TXT = , txt = , tab = , tab.gz = , tab.bz2 = , tab.xz = , txt.gz = , txt.bz2 = , txt.xz = assign(name, read.table(zfile, header = TRUE, as.is = FALSE), envir = tmp_env), CSV = , csv = , csv.gz = , csv.bz2 = , csv.xz = assign(name, read.table(zfile, header = TRUE, sep = \";\", as.is = FALSE), envir = tmp_env), found <- FALSE)\n                }\n                if (found) \n                  break\n            }\n            if (verbose) \n                message(if (!found) \n                  \"*NOT* \", \"found\", domain = NA)\n        }\n        if (found) \n            break\n    }\n    if (!found) {\n        warning(gettextf(\"data set %s not found\", sQuote(name)), domain = NA)\n    }\n    else if (!overwrite) {\n        for (o in ls(envir = tmp_env, all.names = TRUE)) {\n            if (exists(o, envir = envir, inherits = FALSE)) \n                warning(gettextf(\"an object named %s already exists and will not be overwritten\", sQuote(o)))\n            else assign(o, get(o, envir = tmp_env, inherits = FALSE), envir = envir)\n        }\n        rm(tmp_env)\n    }\n}",     "invisible(names)"), check.names = TRUE, fix.empty.names = TRUE,     stringsAsFactors = TRUE): arguments imply differing number of rows: 1, 0, 2, 13